chore: Extract common_handlers module into separate file
Having it in separate file might be a bit cleaner.
